Rename sticky resolver to autolinking resolver
