+++ packages/@expo/cli/src/start/server/metro/withMetroMultiPlatform.ts
@@ -18,13 +18,13 @@ import fs from 'fs';
 import path from 'path';
 import resolveFrom from 'resolve-from';
 
+import {
+  createAutolinkingModuleResolverInput,
+  createAutolinkingModuleResolver,
+  AutolinkingModuleResolverInput,
+} from './createExpoAutolinkingResolver';
 import { createFallbackModuleResolver } from './createExpoFallbackResolver';
 import { createFastResolver, FailedToResolvePathError } from './createExpoMetroResolver';
-import {
-  createStickyModuleResolverInput,
-  createStickyModuleResolver,
-  StickyModuleResolverInput,
-} from './createExpoStickyResolver';
 import { isNodeExternal, shouldCreateVirtualCanary, shouldCreateVirtualShim } from './externals';
 import { isFailedToResolveNameError, isFailedToResolvePathError } from './metroErrors';
 import { getMetroBundlerWithVirtualModules } from './metroVirtualModules';
@@ -167,7 +167,7 @@ export function withExtendedResolver(
   config: ConfigT,
   {
     tsconfig,
-    stickyModuleResolverInput,
+    autolinkingModuleResolverInput,
     isTsconfigPathsEnabled,
     isFastResolverEnabled,
     isExporting,
@@ -176,7 +176,7 @@ export function withExtendedResolver(
     getMetroBundler,
   }: {
     tsconfig: TsConfigPaths | null;
-    stickyModuleResolverInput?: StickyModuleResolverInput;
+    autolinkingModuleResolverInput?: AutolinkingModuleResolverInput;
     isTsconfigPathsEnabled?: boolean;
     isFastResolverEnabled?: boolean;
     isExporting?: boolean;
@@ -193,9 +193,6 @@ export function withExtendedResolver(
   }
   if (isFastResolverEnabled) {
     Log.log(chalk.dim`Fast resolver is enabled.`);
-  }
-  if (stickyModuleResolverInput) {
-    Log.log(chalk.dim`Sticky resolver is enabled.`);
   }
 
   const defaultResolver = metroResolver;
@@ -639,7 +636,7 @@ export function withExtendedResolver(
       return null;
     },
 
-    createStickyModuleResolver(stickyModuleResolverInput, {
+    createAutolinkingModuleResolver(autolinkingModuleResolverInput, {
       getStrictResolver,
     }),
 
@@ -852,7 +849,7 @@ export async function withMetroMultiPlatformAsync(
     exp,
     platformBundlers,
     isTsconfigPathsEnabled,
-    isStickyResolverEnabled,
+    isAutolinkingResolverEnabled,
     isFastResolverEnabled,
     isExporting,
     isReactCanaryEnabled,
@@ -864,7 +861,7 @@ export async function withMetroMultiPlatformAsync(
     exp: ExpoConfig;
     isTsconfigPathsEnabled: boolean;
     platformBundlers: PlatformBundlers;
-    isStickyResolverEnabled?: boolean;
+    isAutolinkingResolverEnabled?: boolean;
     isFastResolverEnabled?: boolean;
     isExporting?: boolean;
     isReactCanaryEnabled: boolean;
@@ -932,16 +929,16 @@ export async function withMetroMultiPlatformAsync(
 
   config = withWebPolyfills(config, { getMetroBundler });
 
-  let stickyModuleResolverInput: StickyModuleResolverInput | undefined;
-  if (isStickyResolverEnabled) {
-    stickyModuleResolverInput = await createStickyModuleResolverInput({
+  let autolinkingModuleResolverInput: AutolinkingModuleResolverInput | undefined;
+  if (isAutolinkingResolverEnabled) {
+    autolinkingModuleResolverInput = await createAutolinkingModuleResolverInput({
       platforms: expoConfigPlatforms,
       projectRoot,
     });
   }
 
   return withExtendedResolver(config, {
-    stickyModuleResolverInput,
+    autolinkingModuleResolverInput,
     tsconfig,
     isExporting,
     isTsconfigPathsEnabled,
